# [Fix #468] Fix false positives for `Performance/BigDecimalWithNumericArgument`