--- a/changelog/fix_false_positives_for_performance_big_decimal_with_numeric_argument.md
+++ b/changelog/fix_false_positives_for_performance_big_decimal_with_numeric_argument.md
@@ -1,0 +1,1 @@
+* [#468](https://github.com/rubocop/rubocop-performance/issues/468): Fix false positives for `Performance/BigDecimalWithNumericArgument` when using float argument for `BigDecimal`. ([@koic][])